Fix the build of test_app_server on 64bit
So many Jamfiles to search through...runs also, but there
are lots of graphical glitches

Reviewed-on: https://review.haiku-os.org/c/haiku/+/1753

+++ src/tests/kits/interface/layout/widget_layout_test/tests/BoxTest.cpp
@@ -122,7 +122,7 @@ BoxTest::ActivateTest(View* controls)
 	fBorderStyleRadioGroup->AddButton(button->GetRadioButton());
 
 	// default to no border
-	fBorderStyleRadioGroup->SelectButton(0L);
+	fBorderStyleRadioGroup->SelectButton((int32)0);
 
 	// spacing
 	group->AddChild(new VStrut(10));
@@ -148,7 +148,7 @@ BoxTest::ActivateTest(View* controls)
 	fLabelRadioGroup->AddButton(button->GetRadioButton());
 
 	// default to no border
-	fLabelRadioGroup->SelectButton(0L);
+	fLabelRadioGroup->SelectButton((int32)0);
 
 	// spacing
 	group->AddChild(new VStrut(10));
